# Improve numba DimShuffle compile time