@aseyboldt 2022-12-11
I'm still trying to understand numba compile times better, but it seems in this case the transpose had quite some impact. And most of the time the transpose in DimShuffle is just a no-op (and we know it is), so we can remove it in those cases.
I also changed the code for the final shape so it involves fewer function calls that need to be analysed and typed.